backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "boto3>=1.40.52",
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "numpy==1.26.1",
    "openai>=2.3.0",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "rq>=2.6.0",
    "sentence-transformers==2.2.2",
    "torch==2.1.0",
    "torchaudio==2.1.0",
    "torchvision==0.16.0",
    "tqdm>=4.67.1",
    "transformers==4.35.0",
]
